refactor persistence out into StorageEmulator
